removed circular dependencies
